Move `SolverPlugin` into its own module
